backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
pgvector
python-dotenv
pdfplumber
google-genai
langchain
langchain-google-genai
langchain-community